fix(ingest): harden SEP-41 metadata surface (#587)
* fix(ingest): prevent malicious SEP-41 token from wedging ledger persistence

  SEP-41 defines decimals() -> u32, but contract_tokens.decimals was
  SMALLINT with the insert cast $8::smallint[]. A contract returning
  a decimals value above 32767 would fail pgx encoding, roll back
  the atomic ledger-persistence transaction, and prevent the ingest
  cursor from advancing past the affected ledger. The same wedge
  class also exists for name()/symbol() strings that exceed Postgres
  TEXT constraints (NUL byte, invalid UTF-8) or grow to unreasonable
  sizes. Reported via HackerOne.

  Widen contract_tokens.decimals to INTEGER (the service is not yet
  in production, so the existing migration is edited in place rather
  than adding a new one). Cap decimals at 70, name at 128 bytes,
  symbol at 32 bytes, and reject non-UTF-8 or NUL-containing strings
  in fetchMetadata. Violations fall through the existing warning path
  so the contract persists with sentinel values (decimals=0, name="",
  symbol="") instead of rolling back the ledger.

* test(ingest): cover FetchSep41Metadata sentinel path for malicious SEP-41 token

  Add TestFetchSep41Metadata_ReturnsSentinelsForMaliciousContract, which
  drives the public entry point with a decimals response above the 70 cap
  and asserts the returned *data.Contract still includes the offending
  contract with sentinel fields (Name="", Symbol="", Decimals=0) and no
  error surfaced. The existing tests only exercised the unexported
  fetchMetadata(), which could not prove that fetchBatch's warning-and-
  continue path is wired through to a persistable placeholder — the actual
  guarantee that stops a malicious token from wedging ledger persistence.

+++ internal/data/contract_tokens.go
@@ -107,7 +107,7 @@ func (m *ContractModel) BatchInsert(ctx context.Context, dbTx pgx.Tx, contracts 
 
 	const query = `
 		INSERT INTO contract_tokens (id, contract_id, type, code, issuer, name, symbol, decimals)
-		SELECT * FROM UNNEST($1::uuid[], $2::text[], $3::text[], $4::text[], $5::text[], $6::text[], $7::text[], $8::smallint[])
+		SELECT * FROM UNNEST($1::uuid[], $2::text[], $3::text[], $4::text[], $5::text[], $6::text[], $7::text[], $8::integer[])
 		ON CONFLICT (contract_id) DO NOTHING
 	`
